Add support for TypeEqual
